+++ packages/solid/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add missing export of `setError` and `setValue` method
+
 ## v0.3.0 (October 30, 2022)
 
 - Change form response object