+++ packages/solid/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
-- Add `clearResponse` method
+- Add `clearResponse` and `hasField` method
 - Add missing export of `setError` and `setValue` method
 
 ## v0.3.0 (October 30, 2022)